Feature/permissions (#44)
* feat:
- updated invites list endpoints response - added organization / user who sent invite information
refactor:
- refactored api module structure and fixed endpoints structure

* feat:
- fixed format, types & lint errors
- fixed schemas errors
- updated tests due to updated in endpoint structure and invites endpoints response

* feat:
- added permissions models, schemas, repository & handler
- added
- added methods for fetching data (select) from db to base crudmixin
- updated endpoints and added rights check
- updated routes and added organization_id query param due to necessity of rights check
- added authorization with bearer token possibility to openapi docs
- added exception handler to app

* fix:
- resolved pr comments
- removed migration with permissions
- removed permissions repository usage
- changed permissions fetch from db to fetch from structure
- moved permissions check from routes to handlers

feat:
- added custom polices for permissions
- added custom polices for permissions
- refactored app configuration

* feat:
- updated tests due to the addition of = permissions checks
- fixed errors

* - resolved rebase conflicts

* fix

* fix:
 - fixed tests

+++ backend/dataforce_studio/handlers/orbits.py
@@ -1,7 +1,9 @@
+from dataforce_studio.handlers.permissions import PermissionsHandler
 from dataforce_studio.infra.db import engine
 from dataforce_studio.infra.exceptions import (
     NotFoundError,
     OrganizationLimitReachedError,
+    ServiceError,
 )
 from dataforce_studio.repositories.orbits import OrbitRepository
 from dataforce_studio.schemas.orbit import (
@@ -13,10 +15,12 @@ from dataforce_studio.schemas.orbit import (
     OrbitUpdate,
     UpdateOrbitMember,
 )
+from dataforce_studio.schemas.permissions import Action, Resource
 
 
 class OrbitHandler:
     __orbits_repository = OrbitRepository(engine)
+    __permissions_handler = PermissionsHandler()
 
     __orbits_limit = 10
 
@@ -31,16 +35,41 @@ class OrbitHandler:
             )
 
     async def create_organization_orbit(
-        self, organization_id: int, orbit: OrbitCreate
+        self, user_id: int, organization_id: int, orbit: OrbitCreate
     ) -> Orbit:
+        await self.__permissions_handler.check_organization_permission(
+            organization_id,
+            user_id,
+            Resource.ORBIT,
+            Action.CREATE,
+        )
+
         await self.check_organization_orbits_limit(organization_id)
 
         return await self.__orbits_repository.create_orbit(organization_id, orbit)
 
-    async def get_organization_orbits(self, organization_id: int) -> list[Orbit]:
+    async def get_organization_orbits(
+        self, user_id: int, organization_id: int
+    ) -> list[Orbit]:
+        await self.__permissions_handler.check_organization_permission(
+            organization_id,
+            user_id,
+            Resource.ORBIT,
+            Action.LIST,
+        )
         return await self.__orbits_repository.get_organization_orbits(organization_id)
 
-    async def get_orbit(self, orbit_id: int) -> OrbitDetails:
+    async def get_orbit(
+        self, user_id: int, organization_id: int, orbit_id: int
+    ) -> OrbitDetails:
+        await self.__permissions_handler.check_orbit_action_access(
+            organization_id,
+            orbit_id,
+            user_id,
+            Resource.ORBIT,
+            Action.READ,
+        )
+
         orbit = await self.__orbits_repository.get_orbit(orbit_id)
 
         if not orbit:
@@ -48,7 +77,17 @@ class OrbitHandler:
 
         return orbit
 
-    async def update_orbit(self, orbit_id: int, orbit: OrbitUpdate) -> Orbit:
+    async def update_orbit(
+        self, user_id: int, organization_id: int, orbit_id: int, orbit: OrbitUpdate
+    ) -> Orbit:
+        await self.__permissions_handler.check_orbit_action_access(
+            organization_id,
+            orbit_id,
+            user_id,
+            Resource.ORBIT,
+            Action.UPDATE,
+        )
+
         orbit_obj = await self.__orbits_repository.update_orbit(orbit_id, orbit)
 
         if not orbit_obj:
@@ -56,16 +95,62 @@ class OrbitHandler:
 
         return orbit_obj
 
-    async def delete_orbit(self, orbit_id: int) -> None:
+    async def delete_orbit(
+        self, user_id: int, organization_id: int, orbit_id: int
+    ) -> None:
+        await self.__permissions_handler.check_orbit_action_access(
+            organization_id,
+            orbit_id,
+            user_id,
+            Resource.ORBIT,
+            Action.DELETE,
+        )
+
         return await self.__orbits_repository.delete_orbit(orbit_id)
 
-    async def get_orbit_members(self, orbit_id: int) -> list[OrbitMember]:
+    async def get_orbit_members(
+        self, user_id: int, organization_id: int, orbit_id: int
+    ) -> list[OrbitMember]:
+        await self.__permissions_handler.check_orbit_action_access(
+            organization_id,
+            orbit_id,
+            user_id,
+            Resource.ORBIT_USER,
+            Action.LIST,
+        )
         return await self.__orbits_repository.get_orbit_members(orbit_id)
 
-    async def create_orbit_member(self, member: OrbitMemberCreate) -> OrbitMember:
+    async def create_orbit_member(
+        self, user_id: int, organization_id: int, member: OrbitMemberCreate
+    ) -> OrbitMember:
+        await self.__permissions_handler.check_orbit_action_access(
+            organization_id,
+            member.orbit_id,
+            user_id,
+            Resource.ORBIT_USER,
+            Action.CREATE,
+        )
+
         return await self.__orbits_repository.create_orbit_member(member)
 
-    async def update_orbit_member(self, member: UpdateOrbitMember) -> OrbitMember:
+    async def update_orbit_member(
+        self,
+        user_id: int,
+        organization_id: int,
+        orbit_id: int,
+        member: UpdateOrbitMember,
+    ) -> OrbitMember:
+        if user_id == member.id:
+            raise ServiceError("You can not update your own data.")
+
+        await self.__permissions_handler.check_orbit_action_access(
+            organization_id,
+            orbit_id,
+            user_id,
+            Resource.ORBIT_USER,
+            Action.UPDATE,
+        )
+
         member_obj = await self.__orbits_repository.update_orbit_member(member)
 
         if not member_obj:
@@ -73,5 +158,17 @@ class OrbitHandler:
 
         return member_obj
 
-    async def delete_orbit_member(self, member_id: int) -> None:
+    async def delete_orbit_member(
+        self, user_id: int, organization_id: int, orbit_id: int, member_id: int
+    ) -> None:
+        if user_id == member_id:
+            raise ServiceError("You can not remove yourself from orbit.")
+
+        await self.__permissions_handler.check_orbit_action_access(
+            organization_id,
+            orbit_id,
+            user_id,
+            Resource.ORBIT_USER,
+            Action.DELETE,
+        )
         return await self.__orbits_repository.delete_orbit_member(member_id)
